% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenbook-package.R
\docType{package}
\name{greenbook-package}
\alias{greenbook}
\alias{greenbook-package}
\title{greenbook: HM Treasury Green Book Cost-Benefit Analysis Primitives}
\description{
Implements cost-benefit analysis primitives from HM Treasury Green Book
guidance: the kinked Social Time Preference Rate (STPR), discount factors,
net present value, equivalent annual cost, and GDP-deflator rebasing.
Pure computation, no network access. Bundled parameter tables carry
vintage metadata for reproducibility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/charlescoverdale/greenbook}
  \item Report bugs at \url{https://github.com/charlescoverdale/greenbook/issues}
}

}
\author{
\strong{Maintainer}: Charles Coverdale \email{charlesfcoverdale@gmail.com}

}
\keyword{internal}
