% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenbook_aid.R
\name{greenbook_aid}
\alias{greenbook_aid}
\title{Function to retrieve customized U.S. troop deployment data}
\usage{
greenbook_aid(
  type = NA,
  recipient = NA,
  startyear,
  endyear,
  value = NULL,
  format = "long"
)
}
\arguments{
\item{type}{The type of aid to return. Values can include economic, military, or all.}

\item{recipient}{The recipient of the aid. Can include Correlates of War (COW) numerical country codes or ISO3C character country codes.}

\item{startyear}{The first year for the series}

\item{endyear}{The last year for the series}

\item{value}{Indicates whether to return constant or current US dollar values for the observed aid amounts.}

\item{format}{Gives the user the option to return the data in long or wide format.}
}
\value{
\code{greenbook_aid()} returns a data frame containing USAID Overseas Loans and Grants data on foreign economic and military aid.
}
\description{
\code{greenbook_aid()} generates a data set containing USAID Overseas Loans and Grants data on foreign economic and military aid.
}
\examples{

\dontrun{
library(greenbook)

example <- greenbook_aid(recipient = NA, startyear = 1980, endyear = 2015)

head(example)

}

}
\references{
United States Government. 2021. US Overseas Loans and Grants (Greenbook). United States Department of State.
}
\author{
Michael E. Flynn
}
